Move App main component to src/components folder
